Add support for animation part which will be played forever
The possibility if providing a last part, which might be played forever
until interrupted is interesting as it allow for an arbitrary time for
finish the boot process.

One restriction is that the forever mode can only be used as a last
part, or it will never move to the next.

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>