open: Support multiple paths as possible animations to load
One of common use we have in embedded devices is the use of OEM boot
splash. To easily support this, we support a sequence of loading paths.

For example, consider following command line:

,----[ Example command line ]
| $: easysplash open /lib/easysplash/oem /lib/easysplash/animation
`----

In the given example, it tries to load from /lib/easysplash/oem and if
it does not exists then it tries /lib/easysplash/animation.

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>